requirements.txt
pyinstaller==6.12.0
google-api-python-client
viam-sdk
requests